Move LogEntryView to separate file
